+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - feat:  add flag `host` in `up` command
+- feat: add installation script
 
 ## v2.6.0 - 2025-08-27